update code: finish news(post in home page)

+++ app/Http/Controllers/Api/CommentServiceController.php
@@ -17,21 +17,38 @@ class CommentServiceController extends Controller
     public function index(Request $request)
     {
         $lst = $request->all();
-        $offset = Constants::OFFSET;
-        $limit = Constants::LIMIT;
-        if ($request['offset'] != null) {
-            $offset = $lst['offset'];
+        if (array_key_exists('post_id', $lst) || $lst['type_comment'] !== null){
+            $offset = Constants::OFFSET;
+            $limit = Constants::LIMIT;
+            if ($request['offset'] != null) {
+                $offset = $lst['offset'];
+            }
+            if ($request['limit'] !=  null) {
+                $limit = $lst['limit'];
+            }
+            if ($lst['type_comment'] !== 'posts' && $lst['type_comment'] !== 'news') {
+                return [
+                    'success' => false,
+                    'code' => 401,
+                    'message' => trans('message.input_not_right')
+                ];
+            } else if ($lst['type_comment'] === 'posts') {
+                $comments = comment::where('post_id', '=', $lst['post_id'])->where('comment_level', '=', 1)->limit($limit)->offset($offset)->get();
+            } else {
+                $comments = comment::where('news_id', '=', $lst['post_id'])->where('comment_level', '=', 1)->limit($limit)->offset($offset)->get();
+            }
+            $result = [
+                'success' => true,
+                'code' => 200,
+                'data' => $comments
+            ];
+            return response()->json($result);
         }
-        if ($request['limit'] !=  null) {
-            $limit = $lst['limit'];
-        }
-        $comments = comment::where('post_id', '=', $lst['post_id'])->where('comment_level', '=', 1)->limit($limit)->offset($offset)->get();
-        $result = [
-            'success' => true,
-            'code' => 200,
-            'data' => $comments
+        return  [
+            'success' => false,
+            'code' => 400,
+            'message' => trans('message.please_fill_out_the_form')
         ];
-        return response()->json($result);
     }
 
     /**
@@ -52,7 +69,17 @@ class CommentServiceController extends Controller
         }
         if (array_key_exists('post_id', $lst) && array_key_exists('comment_content', $lst)) {
             $comment = new comment;
-            $comment->post_id = $lst['post_id'];
+            if ($lst['type_comment'] !== 'posts' && $lst['type_comment'] !== 'news') {
+                return [
+                    'success' => false,
+                    'code' => 401,
+                    'message' => trans('message.input_not_right')
+                ];
+            } else if ($lst['type_comment'] === 'posts') {
+                $like->post_id = $lst['post_id'];
+            } else {
+                $like->news_id = $lst['post_id'];
+            }
             $comment->user_id = $userId;
             $comment->comment_content = $lst['comment_content'];
             if (array_key_exists('comment_id', $lst) && $lst['comment_id'] != null) {
@@ -106,21 +133,38 @@ class CommentServiceController extends Controller
     public function show(Request $request)
     {
         $lst = $request->all();
-        $offset = Constants::OFFSET;
-        $limit = Constants::LIMIT;
-        if ($request['offset'] != null) {
-            $offset = $lst['offset'];
+        if (array_key_exists('post_id', $lst) || $lst['type_comment'] !== null){
+            $offset = Constants::OFFSET;
+            $limit = Constants::LIMIT;
+            if ($request['offset'] != null) {
+                $offset = $lst['offset'];
+            }
+            if ($request['limit'] !=  null) {
+                $limit = $lst['limit'];
+            }
+            if ($lst['type_comment'] !== 'posts' && $lst['type_comment'] !== 'news') {
+                return [
+                    'success' => false,
+                    'code' => 401,
+                    'message' => trans('message.input_not_right')
+                ];
+            } else if ($lst['type_comment'] === 'posts') {
+                $comments = comment::where('post_id', '=', $lst['post_id'])->where('comment_level', '=', 2)->where('comment_id', '=', $lst['comment_id'])->limit($limit)->offset($offset)->get();
+            } else {
+                $comments = comment::where('news_id', '=', $lst['post_id'])->where('comment_level', '=', 2)->where('comment_id', '=', $lst['comment_id'])->limit($limit)->offset($offset)->get();
+            }
+            $result = [
+                'success' => true,
+                'code' => 200,
+                'data' => $comments
+            ];
+            return response()->json($result);
         }
-        if ($request['limit'] !=  null) {
-            $limit = $lst['limit'];
-        }
-        $comments = comment::where('post_id', '=', $lst['post_id'])->where('comment_level', '=', 2)->where('comment_id', '=', $lst['comment_id'])->limit($limit)->offset($offset)->get();
-        $result = [
-            'success' => true,
-            'code' => 200,
-            'data' => $comments
+        return  [
+            'success' => false,
+            'code' => 400,
+            'message' => trans('message.please_fill_out_the_form')
         ];
-        return response()->json($result);
     }
 
     /**
@@ -152,32 +196,39 @@ class CommentServiceController extends Controller
                 'message' => trans('message.unauthenticate')
           ];
         }
-        $commented = comment::where('post_id', '=', $lst['post_id'])->where('comment_id', '=', $lst['comment_id'])->first();
-        if (!$commented) {
-            return [
-                'success' => false,
-                'code' => 403,
-                'message' => trans('message.not_found')
-            ];
+        if (array_key_exists('post_id', $lst) || $lst['type_comment'] !== null){
+            $commented = comment::where('post_id', '=', $lst['post_id'])->where('comment_id', '=', $lst['comment_id'])->first();
+            if (!$commented) {
+                return [
+                    'success' => false,
+                    'code' => 403,
+                    'message' => trans('message.not_found')
+                ];
+            }
+            $commented->comment_content = $lst['comment_content'];
+            $success = $commented->save();
+            if($success != 1){
+                $result = [
+                    'success' => false,
+                    'code' => 400,
+                    'message'=> trans('message.upate_unsuccess'),
+                    'data' => null
+                ];
+            }else{
+                $result = [
+                    'success' => true,
+                    'code' => 200,
+                    'message'=> trans('message.upate_success'),
+                    'data' => $commented
+                ];
+            }
+            return response()->json($result);
         }
-        $commented->comment_content = $lst['comment_content'];
-        $success = $commented->save();
-        if($success != 1){
-            $result = [
-                  'success' => false,
-                  'code' => 400,
-                  'message'=> trans('message.upate_unsuccess'),
-                  'data' => null
-            ];
-        }else{
-            $result = [
-                  'success' => true,
-                  'code' => 200,
-                  'message'=> trans('message.upate_success'),
-                  'data' => $commented
-            ];
-        }
-        return response()->json($result);
+        return  [
+            'success' => false,
+            'code' => 400,
+            'message' => trans('message.please_fill_out_the_form')
+        ];
     }
 
     /**
@@ -196,31 +247,38 @@ class CommentServiceController extends Controller
                 'message' => trans('message.unauthenticate')
           ];
         }
-        $comment = comment::find($id);
-        if ($comment->user_id != $userId) {
+        if (array_key_exists('post_id', $lst) || $lst['type_comment'] !== null){
+            $comment = comment::find($id);
+            if ($comment->user_id != $userId) {
+                return [
+                    'success' => false,
+                    'code' => 401,
+                    'message' => trans('message.can_not_action')
+                ];
+            }
+            if (!$comment) {
+                return [
+                    'success' => false,
+                    'code' => 403,
+                    'message'=> trans('message.not_found_item')
+                ];
+            }
+            $listCommentChildren = comment::where('comment_id', '=', $id)->get();
+            for ($i = 0; $i < count($listCommentChildren); $i++) {
+                $listCommentChildren[$i]->delete();
+            }
+            $comment->delete();
             return [
-                'success' => false,
-                'code' => 401,
-                'message' => trans('message.can_not_action')
+                'success' => true,
+                'code' => 200,
+                'message'=> trans('message.delete_success'),
+                'data' => $comment
             ];
         }
-        if (!$comment) {
-            return [
-                'success' => false,
-                'code' => 403,
-                'message'=> trans('message.not_found_item')
-            ];
-        }
-        $listCommentChildren = comment::where('comment_id', '=', $id)->get();
-        for ($i = 0; $i < count($listCommentChildren); $i++) {
-            $listCommentChildren[$i]->delete();
-        }
-        $comment->delete();
-        return [
-            'success' => true,
-            'code' => 200,
-            'message'=> trans('message.delete_success'),
-            'data' => $comment
+        return  [
+            'success' => false,
+            'code' => 400,
+            'message' => trans('message.please_fill_out_the_form')
         ];
     }
 }
