+++ app/Http/Controllers/Api/LikeServiceController.php
@@ -94,37 +94,54 @@ class LikeServiceController extends Controller
                 'message' => trans('message.unauthenticate')
           ];
         }
-        $liked = like::where('id_post', '=' , $lst['id_post'])->where('id_account_like', '=', $userId)->first();
+        if (array_key_exists('post_id', $lst) || $lst['type_comment'] !== null){
+            $liked = like::where('id_post', '=' , $lst['id_post'])->where('id_account_like', '=', $userId)->first();
             
-        if ($liked) {
-            return [
-                'success' => false,
-                'code' => 401,
-                'message' => trans('message.can_not_action')
-            ];
-        }
-        $like = new like;
-        $like->id_post = $lst['id_post'];
-        $like->id_account_like = $userId;
-        
-        $success = $like->save();
+            if ($liked) {
+                return [
+                    'success' => false,
+                    'code' => 401,
+                    'message' => trans('message.can_not_action')
+                ];
+            }
+            $like = new like;
+            if ($lst['type_like'] !== 'posts' && $lst['type_like'] !== 'news') {
+                return [
+                    'success' => false,
+                    'code' => 401,
+                    'message' => trans('message.input_not_right')
+                ];
+            } else if ($lst['type_comment'] === 'posts') {
+                $like->id_post = $lst['id_post'];
+            } else {
+                $like->id_news = $lst['id_post'];
+            }
+            $like->id_account_like = $userId;
+            
+            $success = $like->save();
 
-        if($success != 1){
-            $result = [
-                  'success' => false,
-                  'code' => 400,
-                  'message'=> trans('message.add_unsuccess'),
-                  'data' => null
-            ];
-        } else {
-            $result = [
-                  'success' => true,
-                  'code' => 200,
-                  'message'=> trans('message.add_success'),
-                  'data' => $like
-            ];
+            if($success != 1){
+                $result = [
+                    'success' => false,
+                    'code' => 400,
+                    'message'=> trans('message.add_unsuccess'),
+                    'data' => null
+                ];
+            } else {
+                $result = [
+                    'success' => true,
+                    'code' => 200,
+                    'message'=> trans('message.add_success'),
+                    'data' => $like
+                ];
+            }
+            return response()->json($result);
         }
-        return response()->json($result);
+        return  [
+            'success' => false,
+            'code' => 400,
+            'message' => trans('message.please_fill_out_the_form')
+        ];
     }
 
     public function unlike (Request $request) {
@@ -137,20 +154,27 @@ class LikeServiceController extends Controller
                 'message' => trans('message.unauthenticate')
           ];
         }
-        $liked = like::where('id_post', '=' , $lst['id_post'])->where('id_account_like', '=', $userId)->first();
-        if (!$liked) {
+        if (array_key_exists('post_id', $lst) || $lst['type_comment'] !== null){
+            $liked = like::where('id_post', '=' , $lst['id_post'])->where('id_account_like', '=', $userId)->first();
+            if (!$liked) {
+                return [
+                    'success' => false,
+                    'code' => 401,
+                    'message' => trans('message.can_not_action')
+                ];
+            }
+            $liked->delete();
             return [
-                'success' => false,
-                'code' => 401,
-                'message' => trans('message.can_not_action')
+                'success' => true,
+                'code' => 200,
+                'message'=> trans('message.delete_success'),
+                'data' => $liked
             ];
         }
-        $liked->delete();
-        return [
-            'success' => true,
-            'code' => 200,
-            'message'=> trans('message.delete_success'),
-            'data' => $liked
+        return  [
+            'success' => false,
+            'code' => 400,
+            'message' => trans('message.please_fill_out_the_form')
         ];
     }
 }
